Remove server.js.
Server.js is now provided by top level repo "mini-components" so this
is not needed. In the end main reason to have this is to enable testing
and providing server in top level repo makes more sense so that all
test code and pages can be served by single server.